Add contact page and story page, finished about pg
Style remains mostly unchanged, although a new
ruleset was added for the footer. There is now the
dungeon scrap mark on the bottom of every page.

All links in the main pages are live.

Button in main folder index file still dead. This
will be styled at a later date, and JS will be
added for some funnies until launch date, when
button will be made live and link to the main
gameplay page.