Add main.js file and built early functions
This is the first chunk of JS. I've started adding
the id selectors and early objects to compose the
skeleton of the game. PC constructor is
prototyped.

A Start Game button has been added to document and
styled. Button is designed to vanish on game start
and reappear on game over.